requirements.txt
telethon
python-dotenv
pandas
sqlalchemy
psycopg2-binary
dbt
dbt-core 
dbt-postgres